questions finished

+++ app/Http/Controllers/QuestionsController.php
@@ -1,13 +1,17 @@
 <?php
 
-namespace App\Http\Controllers;
 
+namespace App\Http\Controllers;
 use App\Question;
 use Illuminate\Http\Request;
 use App\Http\Requests\AskQuestionRequest;
 
+
 class QuestionsController extends Controller
 {
+    public function __construct() {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -15,19 +19,9 @@ class QuestionsController extends Controller
      */
     public function index()
     {
-        //debug
-//        \DB::enableQueryLog();
-//        $questions = Question::with('user')->latest()->paginate(4);
-//        view('questions.index', compact('questions'))->render();
-//        dd(\DB::getQueryLog());
-
-        $questions = Question::with('user')->latest()->paginate(4);
+        $questions = Question::with('user')->latest()->paginate(10);
         return view('questions.index', compact('questions'));
-
-
-
     }
-
     /**
      * Show the form for creating a new resource.
      *
@@ -36,9 +30,8 @@ class QuestionsController extends Controller
     public function create()
     {
         $question = new Question();
-        return view('questions.create',compact('question'));
+        return view('questions.create', compact('question'));
     }
-
     /**
      * Store a newly created resource in storage.
      *
@@ -48,10 +41,8 @@ class QuestionsController extends Controller
     public function store(AskQuestionRequest $request)
     {
         $request->user()->questions()->create($request->only('title', 'body'));
-
-        return redirect()->route('questions.index')->with('success', 'Question submitted');
+        return redirect()->route('questions.index')->with('success', "Your question has been submitted");
     }
-
     /**
      * Display the specified resource.
      *
@@ -60,9 +51,9 @@ class QuestionsController extends Controller
      */
     public function show(Question $question)
     {
-        return view("questions.show", compact('question'));
+        $question->increment('views');
+        return view('questions.show', compact('question'));
     }
-
     /**
      * Show the form for editing the specified resource.
      *
@@ -71,9 +62,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        $this->authorize("update", $question);
         return view("questions.edit", compact('question'));
     }
-
     /**
      * Update the specified resource in storage.
      *
@@ -83,11 +74,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
-        $question->update($request->only('title','body'));
-
-        return redirect('/questions')->with('success','Questions Updated successfuly');
+        $this->authorize("update", $question);
+        $question->update($request->only('title', 'body'));
+        return redirect('/questions')->with('success', "Your question has been updated.");
     }
-
     /**
      * Remove the specified resource from storage.
      *
@@ -96,8 +86,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        $this->authorize("delete", $question);
         $question->delete();
-
-        return redirect('/questions')->with('success','Question deleted');
+        return redirect('/questions')->with('success', "Your question has been deleted.");
     }
 }
+
+?>
